+++ modules/analysis-common/src/main/java/org/elasticsearch/analysis/common/ApostropheFilterFactory.java
@@ -16,30 +16,27 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.elasticsearch.index.analysis;
+package org.elasticsearch.analysis.common;
 
 import org.apache.lucene.analysis.TokenStream;
-import org.apache.lucene.analysis.miscellaneous.ScandinavianFoldingFilter;
+import org.apache.lucene.analysis.tr.ApostropheFilter;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.env.Environment;
 import org.elasticsearch.index.IndexSettings;
+import org.elasticsearch.index.analysis.AbstractTokenFilterFactory;
 
 /**
- * Factory for {@link ScandinavianFoldingFilter}
+ * Factory for {@link ApostropheFilter}
  */
-public class ScandinavianFoldingFilterFactory extends AbstractTokenFilterFactory implements MultiTermAwareComponent {
+public class ApostropheFilterFactory extends AbstractTokenFilterFactory {
 
-    public ScandinavianFoldingFilterFactory(IndexSettings indexSettings, Environment environment, String name, Settings settings) {
+    ApostropheFilterFactory(IndexSettings indexSettings, Environment environment, String name, Settings settings) {
         super(indexSettings, name, settings);
     }
 
     @Override
     public TokenStream create(TokenStream tokenStream) {
-        return new ScandinavianFoldingFilter(tokenStream);
+        return new ApostropheFilter(tokenStream);
     }
 
-    @Override
-    public Object getMultiTermComponent() {
-        return this;
-    }
 }
