+++ modules/analysis-common/src/main/java/org/elasticsearch/analysis/common/CommonAnalysisPlugin.java
@@ -67,7 +67,6 @@ import org.apache.lucene.analysis.standard.ClassicFilter;
 import org.apache.lucene.analysis.tr.ApostropheFilter;
 import org.apache.lucene.analysis.util.ElisionFilter;
 import org.elasticsearch.index.analysis.CharFilterFactory;
-import org.elasticsearch.index.analysis.DelimitedPayloadTokenFilterFactory;
 import org.elasticsearch.index.analysis.HtmlStripCharFilterFactory;
 import org.elasticsearch.index.analysis.PreConfiguredCharFilter;
 import org.elasticsearch.index.analysis.PreConfiguredTokenFilter;
@@ -92,44 +91,53 @@ public class CommonAnalysisPlugin extends Plugin implements AnalysisPlugin {
     @Override
     public Map<String, AnalysisProvider<TokenFilterFactory>> getTokenFilters() {
         Map<String, AnalysisProvider<TokenFilterFactory>> filters = new TreeMap<>();
-        filters.put("asciifolding", ASCIIFoldingTokenFilterFactory::new);
-        filters.put("keyword_marker", requriesAnalysisSettings(KeywordMarkerTokenFilterFactory::new));
-        filters.put("porter_stem", PorterStemTokenFilterFactory::new);
-        filters.put("snowball", SnowballTokenFilterFactory::new);
-        filters.put("trim", TrimTokenFilterFactory::new);
-        filters.put("word_delimiter", WordDelimiterTokenFilterFactory::new);
-        filters.put("word_delimiter_graph", WordDelimiterGraphTokenFilterFactory::new);
-        filters.put("unique", UniqueTokenFilterFactory::new);
-        filters.put("flatten_graph", FlattenGraphTokenFilterFactory::new);
-        filters.put("length", LengthTokenFilterFactory::new);
-        filters.put("lowercase", LowerCaseTokenFilterFactory::new);
-        filters.put("uppercase", UpperCaseTokenFilterFactory::new);
-        filters.put("nGram", NGramTokenFilterFactory::new);
-        filters.put("ngram", NGramTokenFilterFactory::new);
-        filters.put("edgeNGram", EdgeNGramTokenFilterFactory::new);
-        filters.put("edge_ngram", EdgeNGramTokenFilterFactory::new);
-        filters.put("stemmer", StemmerTokenFilterFactory::new);
-        filters.put("stemmer_override", requriesAnalysisSettings(StemmerOverrideTokenFilterFactory::new));
-        filters.put("kstem", KStemTokenFilterFactory::new);
-        filters.put("dictionary_decompounder", requriesAnalysisSettings(DictionaryCompoundWordTokenFilterFactory::new));
-        filters.put("hyphenation_decompounder", requriesAnalysisSettings(HyphenationCompoundWordTokenFilterFactory::new));
-        filters.put("reverse", ReverseTokenFilterFactory::new);
-        filters.put("elision", ElisionTokenFilterFactory::new);
-        filters.put("truncate", requriesAnalysisSettings(TruncateTokenFilterFactory::new));
-        filters.put("limit", LimitTokenCountFilterFactory::new);
-        filters.put("common_grams", requriesAnalysisSettings(CommonGramsTokenFilterFactory::new));
-        filters.put("pattern_replace", requriesAnalysisSettings(PatternReplaceTokenFilterFactory::new));
-        filters.put("pattern_capture", requriesAnalysisSettings(PatternCaptureGroupTokenFilterFactory::new));
+        filters.put("apostrophe", ApostropheFilterFactory::new);
         filters.put("arabic_normalization", ArabicNormalizationFilterFactory::new);
+        filters.put("asciifolding", ASCIIFoldingTokenFilterFactory::new);
+        filters.put("cjk_bigram", CJKBigramFilterFactory::new);
+        filters.put("cjk_width", CJKWidthFilterFactory::new);
+        filters.put("classic", ClassicFilterFactory::new);
+        filters.put("common_grams", requriesAnalysisSettings(CommonGramsTokenFilterFactory::new));
+        filters.put("decimal_digit", DecimalDigitFilterFactory::new);
+        filters.put("delimited_payload_filter", DelimitedPayloadTokenFilterFactory::new);
+        filters.put("dictionary_decompounder", requriesAnalysisSettings(DictionaryCompoundWordTokenFilterFactory::new));
+        filters.put("edge_ngram", EdgeNGramTokenFilterFactory::new);
+        filters.put("edgeNGram", EdgeNGramTokenFilterFactory::new);
+        filters.put("elision", ElisionTokenFilterFactory::new);
+        filters.put("fingerprint", FingerprintTokenFilterFactory::new);
+        filters.put("flatten_graph", FlattenGraphTokenFilterFactory::new);
         filters.put("german_normalization", GermanNormalizationFilterFactory::new);
         filters.put("hindi_normalization", HindiNormalizationFilterFactory::new);
+        filters.put("hyphenation_decompounder", requriesAnalysisSettings(HyphenationCompoundWordTokenFilterFactory::new));
         filters.put("indic_normalization", IndicNormalizationFilterFactory::new);
+        filters.put("keep", requriesAnalysisSettings(KeepWordFilterFactory::new));
+        filters.put("keep_types", requriesAnalysisSettings(KeepTypesFilterFactory::new));
+        filters.put("keyword_marker", requriesAnalysisSettings(KeywordMarkerTokenFilterFactory::new));
+        filters.put("kstem", KStemTokenFilterFactory::new);
+        filters.put("length", LengthTokenFilterFactory::new);
+        filters.put("limit", LimitTokenCountFilterFactory::new);
+        filters.put("lowercase", LowerCaseTokenFilterFactory::new);
+        filters.put("min_hash", MinHashTokenFilterFactory::new);
+        filters.put("ngram", NGramTokenFilterFactory::new);
+        filters.put("nGram", NGramTokenFilterFactory::new);
+        filters.put("pattern_capture", requriesAnalysisSettings(PatternCaptureGroupTokenFilterFactory::new));
+        filters.put("pattern_replace", requriesAnalysisSettings(PatternReplaceTokenFilterFactory::new));
         filters.put("persian_normalization", PersianNormalizationFilterFactory::new);
+        filters.put("porter_stem", PorterStemTokenFilterFactory::new);
+        filters.put("reverse", ReverseTokenFilterFactory::new);
+        filters.put("scandinavian_folding", ScandinavianFoldingFilterFactory::new);
         filters.put("scandinavian_normalization", ScandinavianNormalizationFilterFactory::new);
         filters.put("serbian_normalization", SerbianNormalizationFilterFactory::new);
+        filters.put("snowball", SnowballTokenFilterFactory::new);
         filters.put("sorani_normalization", SoraniNormalizationFilterFactory::new);
-        filters.put("cjk_width", CJKWidthFilterFactory::new);
-        filters.put("cjk_bigram", CJKBigramFilterFactory::new);
+        filters.put("stemmer_override", requriesAnalysisSettings(StemmerOverrideTokenFilterFactory::new));
+        filters.put("stemmer", StemmerTokenFilterFactory::new);
+        filters.put("trim", TrimTokenFilterFactory::new);
+        filters.put("truncate", requriesAnalysisSettings(TruncateTokenFilterFactory::new));
+        filters.put("unique", UniqueTokenFilterFactory::new);
+        filters.put("uppercase", UpperCaseTokenFilterFactory::new);
+        filters.put("word_delimiter_graph", WordDelimiterGraphTokenFilterFactory::new);
+        filters.put("word_delimiter", WordDelimiterTokenFilterFactory::new);
         return filters;
     }
