+++ modules/analysis-common/src/test/java/org/elasticsearch/analysis/common/CommonAnalysisFactoryTests.java
@@ -22,7 +22,6 @@ package org.elasticsearch.analysis.common;
 import org.apache.lucene.analysis.charfilter.HTMLStripCharFilterFactory;
 import org.apache.lucene.analysis.en.PorterStemFilterFactory;
 import org.apache.lucene.analysis.miscellaneous.LimitTokenCountFilterFactory;
-import org.apache.lucene.analysis.payloads.DelimitedPayloadTokenFilterFactory;
 import org.apache.lucene.analysis.reverse.ReverseStringFilterFactory;
 import org.apache.lucene.analysis.snowball.SnowballPorterFilterFactory;
 import org.elasticsearch.index.analysis.HtmlStripCharFilterFactory;
@@ -117,6 +116,15 @@ public class CommonAnalysisFactoryTests extends AnalysisFactoryTestCase {
         filters.put("soraninormalization", SoraniNormalizationFilterFactory.class);
         filters.put("cjkwidth", CJKWidthFilterFactory.class);
         filters.put("cjkbigram", CJKBigramFilterFactory.class);
+        filters.put("delimitedpayload", DelimitedPayloadTokenFilterFactory.class);
+        filters.put("keepword", KeepWordFilterFactory.class);
+        filters.put("type", KeepTypesFilterFactory.class);
+        filters.put("classic", ClassicFilterFactory.class);
+        filters.put("apostrophe", ApostropheFilterFactory.class);
+        filters.put("decimaldigit", DecimalDigitFilterFactory.class);
+        filters.put("fingerprint", FingerprintTokenFilterFactory.class);
+        filters.put("minhash", MinHashTokenFilterFactory.class);
+        filters.put("scandinavianfolding", ScandinavianFoldingFilterFactory.class);
         return filters;
     }
 
@@ -155,7 +163,7 @@ public class CommonAnalysisFactoryTests extends AnalysisFactoryTestCase {
         filters.put("common_grams", null);
         filters.put("czech_stem", null);
         filters.put("decimal_digit", null);
-        filters.put("delimited_payload_filter", DelimitedPayloadTokenFilterFactory.class);
+        filters.put("delimited_payload_filter", org.apache.lucene.analysis.payloads.DelimitedPayloadTokenFilterFactory.class);
         filters.put("dutch_stem", SnowballPorterFilterFactory.class);
         filters.put("edge_ngram", null);
         filters.put("edgeNGram", null);
