+++ modules/analysis-common/src/test/resources/org/elasticsearch/analysis/common/keep_analysis.json
